backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
lightgbm==4.0.0
xgboost==2.0.0
joblib==1.3.2
osmnx==1.6.0
networkx==3.1
openrouteservice==2.4.0
shap==0.42.1
matplotlib==3.7.2
plotly==5.15.0
folium==0.14.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.20
alembic==1.12.0
pydantic==2.0.3
pydantic-settings==2.0.3
python-dateutil==2.8.2
python-multipart==0.0.6
